+++ documentation/RELEASE_NOTES.md
@@ -3,7 +3,7 @@
 - Fixed issue with put into PvObjectQueue when timeout is given
 - Updated default value for PvTimeStamp userTag field to 0
 - Area Detector Simulator enhancements:
-  - added ability to load images from HDF5 files
+  - added ability to load images from HDF5 files (either compressed or uncompressed)
 - Conda/pip package dependencies: 
   - EPICS BASE = 7.0.7
   - BOOST      = 1.78.0